Dev css changes (#98)
* auto scaling font size on config dialog for high res monitors

* customizable gap between chat entries

* customizable font size and menu size for high res monitors

* reorganized config navigation

* reorganized config pages

* updated ui engine

* updated sharlayan resources

* switch to typescript and scss

+++ Gobchat.App/Core/Util/ArchiveUnpacker.cs
@@ -24,7 +24,7 @@ namespace Gobchat.Core.Util
 
         public bool DeleteArchiveOnCompletion { get; set; } = true;
 
-        public bool DeleteOutputFolderOnFail { get; set; } = false;
+        public bool DeleteOutputFolderOnFail { get; set; } = true;
 
         public ArchiveUnpacker(string archivePath, string outputFolderPath)
         {
@@ -47,7 +47,8 @@ namespace Gobchat.Core.Util
             catch (ExtractionFailedException ex)
             {
                 progressMonitor.Log(StringFormat.Format(Resources.GeneralErrorOccured, ex.Message));
-                DeleteExtractedData(progressMonitor);
+                if (DeleteOutputFolderOnFail)
+                    DeleteExtractedData(progressMonitor);
                 DeleteArchive(progressMonitor);
                 throw;
             }
@@ -61,7 +62,8 @@ namespace Gobchat.Core.Util
                     return Result.Completed;
 
                 case ArchiveUnpackerHelper.ExtractionResult.Canceled:
-                    DeleteExtractedData(progressMonitor);
+                    if (DeleteOutputFolderOnFail)
+                        DeleteExtractedData(progressMonitor);
                     return Result.Canceled;
             }
 
@@ -76,9 +78,8 @@ namespace Gobchat.Core.Util
 
         private void DeleteExtractedData(IProgressMonitor progressMonitor)
         {
-            progressMonitor.Log(Resources.Core_Util_ArchiveUnpacker_DeleteIncomplete);
-            if (DeleteOutputFolderOnFail)
-                Directory.Delete(OutputFolderPath, true);
+            progressMonitor.Log(Resources.Core_Util_ArchiveUnpacker_DeleteIncomplete);            
+            Directory.Delete(OutputFolderPath, true);
         }
 
         public enum Result
